fix weechat persistence
This is still a hack.

Something is wrong with permissions upstream. They document 777 as the
solution but that should never be a solution. I think I want to explore
the container build, especially since the permissions error is reported
on github issues #10. They provided an unrelated fix to UID inside the
container, but it didn't help about 777.