fix(linkwarden): discard secrets for now
Environment= keys in quadlets are populated before Secret= is added to
the container environment, I think. This may be why the podman secret
for the db pw is empty in the database url string.

EnvironmentFile will probably have the same issue, so right now my poc
system is using the password in plaintext. plsfix