fix(quadlets): 1 .volume file per volume
I saw an example that added multiple `VolumeName` entries in a single
`.volume` file but upon revisiting I don't see that working as intended.

It does seem that podman/systemd will create the volumes defined in
`.container` if you use the `Volume=foobar:/foo/bar` syntax, without any
need for the `.volume` file. This may result in unnamed volumes, but
still work, as if `VolumeName` was not defined. Further investigation
required.